Basic, working SPA setup
