+++ .gitignore
@@ -20,6 +20,7 @@ static/js/cdr-admin.js
 static/css/cdr_admin.css
 static/js/cdr-access.js
 static/css/cdr_access.css
+static/js/vue-access.js
 *.log*
 *.iml
 dump.rdb
