+++ static/js/vue-cdr-access/.browserslistrc
@@ -1,0 +1,2 @@
+> 1%
+last 2 versions
