+++ static/js/vue-cdr-access/babel.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  presets: [
+    '@vue/app'
+  ]
+}
